Add option to select kernel type
This option enables to set kernel type in force.

+++ tools/sdk_import.sh
@@ -13,12 +13,13 @@ SCRIPT_DIR=`dirname "${SCRIPT_NAME}"`
 # Help message
 function show_help
 {
-	echo "Usage: $0 [OPTION...]"
+	echo "Usage: $0 [OPTION...] <SDK export zip file>"
 	echo ""
 	echo "Import SDK export archive into Arduino environment."
 	echo ""
 	echo "  Optional arguments:"
 	echo ""
+	echo "    -k: Force kernel type(release/debug)"
 	echo "    -o: Prebuilt output directory"
 	echo "    -h: Show help (This message)"
 	echo ""
@@ -26,13 +27,18 @@ function show_help
 
 # Option handling
 SDK_DIR="${SCRIPT_DIR}/../Arduino15/packages/SPRESENSE/tools/spresense-sdk/1.0.0"
-while getopts o:h OPT
+FORCE_KERNEL_TYPE=""
+while getopts k:o:h OPT
 do
 	case $OPT in
+		'k' ) FORCE_KERNEL_TYPE=$OPTARG ;;
 		'o' ) SDK_DIR=`cd $OPTARG && pwd`;;
 		'h' ) show_help;;
 	esac
 done
+
+# Shift argument position after option
+shift $(($OPTIND - 1))
 
 if [ $# != 1 ]; then
 	echo "Usage: $0 <SDK export file>"
@@ -71,10 +77,18 @@ else
 fi
 
 # Check release/debug kernel config
-if [ "`grep "CONFIG_DEBUG_FEATURES=y" ${EXP_DIR}/nuttx/.config`" != "" ]; then
-	CONFIG_ENABLE_DEBUG=true
+if [ "${FORCE_KERNEL_TYPE}" == "" ]; then
+	if [ "`grep "CONFIG_DEBUG_FEATURES=y" ${EXP_DIR}/nuttx/.config`" != "" ]; then
+		CONFIG_ENABLE_DEBUG=true
+	else
+		CONFIG_ENABLE_DEBUG=false
+	fi
 else
-	CONFIG_ENABLE_DEBUG=false
+	if [ "${FORCE_KERNEL_TYPE}" == "debug" ]; then
+		CONFIG_ENABLE_DEBUG=true
+	else
+		CONFIG_ENABLE_DEBUG=false
+	fi
 fi
 
 # Select kernel config
@@ -150,7 +164,7 @@ do
 done
 
 #If target is subcore, copy prebuilt into debug
-if [ "${SDK_CONF}" == "spresense_sub" ]; then
+if [ "${SDK_CONF}" == "spresense_sub" -a "${FORCE_KERNEL_TYPE}" == "" ]; then
 	DEBUG_KERNEL_DIR=${KERNEL_DIR}/../subcore-debug
 	rm -rf ${DEBUG_KERNEL_DIR}
 	cp -a ${KERNEL_DIR} ${DEBUG_KERNEL_DIR}
